Make create_fs() error handling as bulletproof as possible
Ok so now we're calling udev_settle. What if *that* fails? We're really
getting out on a limb of cascading failures here. But try to just sleep
to give time to free the device and keep going.

If thin_dev.destroy() fails at this point, I don't know what else there
is to do. Emit to the log and add a TODO.

Signed-off-by: Andy Grover <agrover@redhat.com>