Handle if save_fs() fails
Attempt to destroy dangling DM device if save_fs() fails.

Refactor error handling from StratFilesystem::initialize() into
destroy(), so this second calling location also benefits. This involves
creating the StratFilesystem earlier in initialize() so we can call a
method on it. (This seemed like a reasonable thing to do.)

Taking this approach still doesn't completely fix the issue, because it
involves undoing things, and the undoing can also potentially fail. But I
think it is an improvement: we are now TWO extremely unlikely events
-- save_fs() failing and then fs.destroy() failing -- away from an
unsupported state.

fixes #1274

Signed-off-by: Andy Grover <agrover@redhat.com>