fs::initialize and thinpool::create_filesystem both need to settle
Partially Revert "Handle if save_fs() fails".

If any step in creating a fs fails (after mkfs), we need to possibly wait
for udev to finish before we can remove the thindev. Since this can
happen in both StratFilesystem::initialize (if create_fs fails) or in
ThinPool::create_filesystem (if save_fs fails), add a helper fn called
fs_settle() that does this. This is instead of the previous attempt, that
tried to do the settle in destroy(), but this wasn't quite right since
settling is only really needed in creation-related error handling.

Signed-off-by: Andy Grover <agrover@redhat.com>