Make a method to remove specified blockdevs
This method allows reversing the decision of having added blockdevs to
a blockdev manager and having initialized them.

Use the method in CacheTier::add if the devices added to the cache tier
would cause the size of the cache sub-device to exceed its specified limit.
Do the same in CacheTier::new. Change the return type, to accomodate the
possibility of a failure.

Signed-off-by: mulhern <amulhern@redhat.com>