Use proptest instead of quickcheck
There is a discussion of the pros and cons of this linked to from the
quickcheck homepage. Aside from those mentioned, it is just easier to
be fluent using proptest, and it requires less work to generate a broad
range of examples from the test values. With quickcheck the work was
difficult enough that it just wasn't being done in the existing quickcheck
tests.

What is not clear is how valuable the existing quickcheck tests actually
are at this time. All we really know is that in the early days they flushed
out a bunch of bugs and that they continue to pass now.

Signed-off-by: mulhern <amulhern@redhat.com>