Change specification of extend_thin_sub_device
Do not return an error if it can allocate only 0 sectors. This make the
behavior more consistent and easier to deal with. Errors only from failure
to actually get devicemapper to do what is wanted.

Put in checks for a 0 extend amount, as they are now possible. This change
preserves semantics, except for the content of warning messages.

Make error message if amount is 0 precise. Previous error message
might cause reader to infer that the device hat hit some intrinsic and
permanent limit. This log message makes it more clera that the problem is
extrinsic, and might be fixable.

Signed-off-by: mulhern <amulhern@redhat.com>