Implement greedy thinpool allocation (take 2)
Instead of in create_pool and add_datadev, implement greedy extension in
check(). This keeps everything more in one place, for when we look to
reimplement datadev extensions. I have chosen to remove code rendered
inoperative by this change rather than just commenting it out. (We can
pull it back out of history when the time comes.)

Add Backstore::datatier_usable_size() since this needs it.

Signed-off-by: Andy Grover <agrover@redhat.com>