Re-fetch thinpool status if there were thinpool extensions in check()
check() checks overall pool status first, and then may extend meta or
data segments as needed. But, if the extensions brought the pool back
into a good condition, pool status is not updated.

So, if an extension has happened and the previous pool state was bad,
call status() again to possibly update PoolState.

Signed-off-by: Andy Grover <agrover@redhat.com>