Use a similar but shorter condition to stop writing
This loop also stops if the pool or extend states are in their
Initializing or Stopping states. They shouldn't ever be in those states,
but the test should stop writing if they are.

It also stops if the pool has gone into a Failed state. That is obviously
the correct thing to do, should it happen.

Signed-off-by: mulhern <amulhern@redhat.com>