Write a function to map thinpool status to pool state
Use it where appropriate.

Also, make error messages only about current thinpool state and get rid
of the exclamation points. The pool state is recorded in the set_state
method, so it doesn't need to be mentioned in the log message. Exclamation
points are redundant; the log level makes it clear that the entry is
considered important by the writer.

Signed-off-by: mulhern <amulhern@redhat.com>