Get the thin pool status less conditionally
Do this unconditinally, as it is uncertain whether the actions taken
or not taken may have made the thin pool status better or worse.

Set the pool state according to the status, regardless of its value as
the set state method checks if there has been a change (and exits
immediately if there hasn't).

Signed-off-by: mulhern <amulhern@redhat.com>