Add an auxiiliary method to combine getting status and setting state
Generally speaking, it is wrong to miss an opportunity to set the state
if the status would indicate that the state should be changed.

Note that the method is local to check, because, although it would be
better to set the state everywhere, it is impossible to do this in
ThinPool::total_physical_used, because that would require self to be
mutable.

Do not extend the auxiliary method's use to the tests. They seem to be
directed toward a lower level, checks about the DM thin pool device, so that
the use of the bare DM thin pool status method seems more or less correct.

Signed-off-by: mulhern <amulhern@redhat.com>