Remove process_deferred_actions from DbusConnection::register_pool
process_deferred_actions() is a method that processes miscellaneous
actions that arise from interactions on the D-Bus. It doesn't have to do
with a specific pool, so sticking it at the end of
DbusConnection::register_pool is a bit misleading.

Also, in all places where DbusConnection::register_pool is called, it is
not desirable to process deferred actions.

It is called in MaybeDbusSupport::setup_connection when the connection has
first been set up. stratisd is not
multi-threaded, and the connection has not been set up until this point.
Therefore, there can not be any deferred actions to process. So, calling
process_deferred_actions in this context is a null op. It can be skipped.

It is also called transitively by UdevMonitor::handle_events if a new pool
has been constructed in response to a udev event. In this case, there can
be no deferred actions on the previously non-existent pool, so processing
deferred actions can be skipped.

This means that process_deferred_actions is now called in exactly one place,
the correct one, where _D-Bus_ events are handled. The deferred actions
were originally designed with this and only this mechanism in mind.

Signed-off-by: mulhern <amulhern@redhat.com>