Rework Backstore::request()
Do a few things:

Replace an algorithm that loops to discover how much of the data tier
that can be allocated with one that just ASKS how much space is available
and then modifies the single param to alloc() based on what's available.

Align the request UP to a multiple of the modulus instead of down. This
avoids issues where percentage-based requests fall short by some small
number of sectors and then try in vain repeatedly to get the difference
but keep failing.

Align the sectors allocated DOWN to a multiple of the modulus. This is
needed because cap extensions are *not* necessarily aligned to the
modulus, such as when there are a small number of unused sectors in the
cap already, or if remaining free sectors in data_tier is small.

Implement align_up() and align_down() as helper functions, with tests,
since this stuff is a little tricky, and maybe we will find other spots
in the future that need aligning.

Signed-off-by: Andy Grover <agrover@redhat.com>