Reorganise formatting task
Specify the compiler version using Travis.

Move all non-test, i.e., setup stuff, out of the test into Travis.
This is how Travis distinguishes between an actual test failure
and a failure to set up the test properly.

The web display for Travis is made a bit more helpful by this, because
the actual compiler version used is visible.

There is now a consistency that makes sense, since fmt and clippy target
are set up in the same way.

Signed-off-by: mulhern <amulhern@redhat.com>