Remove dbus support
For the two to co-exist we would need to figure out a way for changes
over the varlink interface to be reflected in dbus as dbus interface
stores quite a bit of meta-data outside of the engine.  Changes made
over the dbus interface are correctly shown over the varlink
interface.

Signed-off-by: Tony Asleson <tasleson@redhat.com>