Precondition check for device usage
Check that a block device isn't already in use for the engine.

This works, but isn't ideal as the engine caller needs to validate
block device preconditions instead of the calls themselves. This occurs
as we don't have the entire engine context known when we are executing
code for a specific pool for example.  The following additional approaches
were attempted.

* We cannot pass the engine to the calls that need it as it is already
  borrowed mutably.
* Refactor the interface, so that we call methods on the engine instead of
  dispatching to the individual items contained in the engine, this results
  in the engine knowing too much of the internals of the pool.

Addresses: https://github.com/stratis-storage/stratisd/issues/1291

Signed-off-by: Tony Asleson <tasleson@redhat.com>