Return an error under two more sets of circumstances
If the sigblocks differ but the initialization times agree return an error
There is no good explanation for why this should happen, and no way to
know how to repair.

Return an error if unable to read data at either sigblock location.
This preserves previous behavior and seems more correct to me.

Signed-off-by: mulhern <amulhern@redhat.com>