Updating test to make sure rebuild of the filesystem has extended size
Reorganized a bit because filesystem.check() only updates XFS.  To test
that the MDV is updated property - call pool.check() instead to update
both XFS and the thindev size on the MDV.

Limit the scope of the mutalbe borrow of the pool to allow the subsequent
call to pool.check().

Updated the test name to reflect that it is no longer XFS specific.
It now tests both thindev reconstruction and XFS update when a thindev is
extended.

Simulate reboot by adding teardown and reconstruction of the thindev pool
then verify extended size for the thindev.

Signed-off-by: Todd Gill <tgill@redhat.com>