Downgraded 'Data tier percent used' message from info to debug
The message 'Data tier percent used' was being displayed to the user at the
info level, which is unnecessary for users who don't use up a significant
amount of space.

In this PR, I downgraded the original message to the debug level
and added an info level message which displays the data tier percent used
only when the said percent has reached a critical level.