Make no udev test conditional on STRATIS_DESTRUCTIVE_TEST
This environment variable was originally proposed when udev tests were
first added in dbus test directory. At that time it was desirable to have
some filter to keep the udev tests, which tests udev behavior and
therefore affect the environment, distinct from the dbus tests, which
interact with the sim engine and therefore do not effect the environment.

Now that the udev tests are in a separate directory and run independently,
this environment variable is unnecessary.

Signed-off-by: mulhern <amulhern@redhat.com>