test_udev.py: Correct pool create
When we dynamically create loop back devices we need to ensure that
the devices exist in the udev db before we issue a dbus call to
create the pool.  Otherwise stratisd will be unable to find the
block device and return an error on the pool create.

This correction adds settles after the loop back devices are
created and also adds retries when creating the stratis pool.

Signed-off-by: Tony Asleson <tasleson@redhat.com>