Use types to distinguish between sizes of MDA parts
Have upper-level methods request the size they need to store their data,
rather than the size of the entire MDA region.

This allows to design validate_mda_size() out of existance, because the
metadata module always generates a valid mda size, based on the size that
the caller requires to store variable length metadata.

This also allows better encapsulation. Instead of having to export
MIN_MDA_SECTORS, which ought to be a private detail, and validate_mda_size,
which is now entirely unnecessary, the metadata module only exports
MDADataSize.

Signed-off-by: mulhern <amulhern@redhat.com>