Do not use MIN_MDA_SECTORS in inappropriate contexts
It is used in two places to specify the number of bytes to wipe on
some blockdevs, when the purpose is to wipe the header block which includes
the two signature blocks. The header block is 16 sectors, and the
minimum MDA size is 2032 sectors, so the tests were always wiping enough.
But using MIN_MDA_SECTORS in that context is confusing to someone who
reads the test, so better to use the correct amount.

In one of the two places, it was not part of the test and should have
been unnecessary, so it is removed there.

Signed-off-by: mulhern <amulhern@redhat.com>