Make a simple, private method to parse an MDAHeader buffer
Due to the nature of the values in the MDAHeader, parsing can not return an
error. It is better to make this obvious by omitting the error handling
from the code that does the parsing. The separate method is helpful in
clarifying the separation of the checking validity from the assigning of
the bits to the different fields.

Signed-off-by: mulhern <amulhern@redhat.com>