Change MDAHeader::parse_buf to check used rather than time
If the length of the variable length metadata written is 0, then there
is no variable length metadata to be read. This makes more sense than
checking the date, since it is at least possible that a 0 time could still
correspond to some metadata.

Fix up some tests.  Don't allow any test to write 0 length variable length
metadata.  Do allow a test to have a 0 valued secs value for a timestamp.

Signed-off-by: mulhern <amulhern@redhat.com>