Add a vendor target to the Makefile
To make vendoring experiments a bit more convenient.

Get rid of the uml-graphs item in the PHONY list.
The target went away a while ago, but somehow the uml-graphs item
escaped the cull. This change is _just_ too small and inconsequential to
put in a separate PR, so I'm putting it in here.

Signed-off-by: mulhern <amulhern@redhat.com>