Remove an assertion that is evaluated at compile-time
The compiler does the comparison at run time, verifies that it is true,
and then removes the assertion entirely.

Instead, have a run-time assertion that asserts precisely what is required
for correctness (and what we believe is a logical necessity of the way
our code is constructed and of unsigned computer arithmetic).

Signed-off-by: mulhern <amulhern@redhat.com>