Move three functions out of the StaticHeader impl
These functions are slightly better at the module level, now that a module
exists.

The read function just reads some bytes according to the static header
geometry.

The wipe function just wipes some bytes according to the static header
geometry.

The setup function does is not an instance method and may or may not
actually yield a StaticHeader.