Move device_identifiers from StaticHeader to BDA
Also, improved the header docs.

In this way StaticHeader can be made private to the metadata module and
better encapsulation is achieved.

It does not make much sense to export StaticHeader and all its public
methods from the metadata module in order to make accessible a one-line
convenience method.

This ought also to ease further refactoring of the static header
specific functionality into a separate module.

Signed-off-by: mulhern <amulhern@redhat.com>