Make beta task clippy rather than test-travis
It's a more complete way to look at the future, because clippy must
compile virtually everything in order to run its lints. So, if the
test fails while compiling we know there is something to do, and if the
test fails while linting, we know there is something to do.

Signed-off-by: mulhern <amulhern@redhat.com>