Remove is_stratis_device
It's just a convenience method for accessing the fields of
DevOwnership::Ours if that's the variant found. Put a convenience method
in DevOwnerhip for doing just that, on the principle of least strength.

Eliminate all tests of is_stratis_device.

Where is_stratis_device is used in the source, substitute an exactly
equivalent use of identify() and some other mechanism as appropriate.

For block_evaluate(), choose the textual change which will lead to the least
reformatting by rustfmt.

Don't add more additional methods to DevOwnership. Its overall design is
due for a rework, so additional methods would be overkill.

Signed-off-by: mulhern <amulhern@redhat.com>