Get rid of special ext3 creating method
It's not consistent with the other methods, because a wipefs call was
placed in it.

It is possible that it is entirely unnecessary, and the same tests can
be adequately done using the xfs filesystem create method.

Update string matching on "ext3" to string matching on "xfs" in tests as
appropriate.

Signed-off-by: mulhern <amulhern@redhat.com>