Fix a TODO
It implies that failure to obtain the ID_WWN is always due to some
problem, but it is also possible that the device simply does not have an
ID_WWN. I'm not sure that no ID can be addresed by anything other than
finding a new device, which the user may not be interested in doing.
We should certainly not log a missing ID at the warn level, and it's
not really clear that we should even log it at the info level. If the user
has a lot of such devices, they'll be plenty bored with seeing the same
information whenver stratisd comes up.

Signed-off-by: mulhern <amulhern@redhat.com>