Remove get_all_empty_devices
It is only used in one place, and the code is a bit clearer if it is only
in that one place. Also, it is slightly more efficient, in that it
iterates once over the list of devices obtained from udev, rather than
making that list into a Vec, and then reiterating over the Vec.

Move the first udev context and enumerator that occurs in the same method
into a separate scope, for clarity.

Signed-off-by: mulhern <amulhern@redhat.com>