D-Bus interface changes for encryption implementation
This commit makes use of the crypt.rs module implementation of encryption
by exposing the encryption implementation through the pool interface. The
key_desc parameter has been added on the D-Bus which, when provided,
causes create_pool to encrypt all block devices with the specified key
in the kernel keyring using the key description passed using the D-Bus
call.

Co-Authored-By: mulhern <amulhern@redhat.com>