Move udev-related identifying into udev module
Also, remove signature function. It's misleading, because it makes it
look like stratis actually looked at those properties to determine the
device ownership, which it didn't. Also, it is a refactoring burden.

Add a FIXME on the DevOwnership::Theirs constructor.
Where it is matched against in blockdevmgr::initialize say something truer
to its real meaning.

The empty() function now gone was equivalent to
!is_multipath_member(&device) && is_unclaimed(&device).

Note that the checks in the decide_ownership() method are in a slightly
different order from previous, but I think in an equivalent order.
I did my best to document what I believe may be the ordering constraints.

Note that the identify method identifies a multipath member device as
Theirs which is a very rough classification, since a multipath member that
belongs to Stratis is also ours. I haven't tried to address that.

Signed-off-by: mulhern <amulhern@redhat.com>