Do not call identify() method in find_all() method
The drawback of doing this was confusion due to multiple redundant
checks for the same properties, i.e., checking is_multipath_member and
is_unclaimed both in find_all() and in identify(), and also iterating
twice through all the udev devices, essentially making the function
quadratic instead of linear.

Use the two component methods decide_ownership(), which already handles
checking for multipath membership and unclaimedness via udev, and
DevOwnership::from_udev_ownership() which does the remainder of the
actions in identify.

This is intended to be semantics preserving:

* All devices that don't have a devnode are excluded.
* All devices considered multipath member devices are excluded.
* All devices considered unclaimed are included.
* If there is an error in getting the information, the device is just
ignored.

Fix up the comment that describes the code so that it describes what the
code actually does.

Note: This is a slight change, because the previously existing code would
have excluded from its search a device that appeared to be a Stratis device,
but the current search will not. This would have happened if in the interval
between the first udev enumeration and the second, some devices properly
labeled Stratis had actually appeared.

Signed-off-by: mulhern <amulhern@redhat.com>