Get the ownership designation and the hardware id at the same time
This way, all devices belonging to the udev database are iterated through
just once for each block device passed to udev_initialize, instead of twice.

Also, the missing udev entry possibility is handled in the one place that
the udev devices are iterated over, by returning an error if it can't be
found, as before. Previously, with the lookup of hw_id, it ignored a missing
udev entry, which made sense, because it had sort of proved that one
existed, by finding it at an earlier stage in the method. But, that now
leaves the only possibility of failure as a failure to decode the hw id
properly, so I'll just log that as a warn. This takes care of the TODO
and is slightly better than what the previous code would do: in the unlikely
event of an error decoding the ID_WWN it would return a lossy string.

Allow type_complexity on dev_info, it is an internal method.

Get rid of some obsoleted methods.

Signed-off-by: mulhern <amulhern@redhat.com>