Ensure that the devices found only correspond to the pools constructed
During work on PR https://github.com/stratis-storage/stratisd/pull/1672
intermittent errors were experienced in test_setup. The reported error was:

UUIDs of devices found () did not correspond with UUIDs specified in the
metadata for this group of devices (223b6299d883449e95dff1d2ddaa05c7)

I believe that the cause of this error was that find_all() would find the
devices for both the pool made by the test _and_ for a pool corresponding
to some devices left around from a previous test. Sometimes the devices
corresponding to the test pool were placed first in the result of find_all()
and hence were passed to the setup() function, and the test succeeded.
Other times the device left over was found first and passed to the setup()
function, and so the test failed.

These changes to the test ensure that a problem like that would be detected
immediately if it recurred.

Signed-off-by: mulhern <amulhern@redhat.com>