Remove some unnecessary implicit-hasher clippy allows
I think these were necessary once, because the methods were exposed in the
public interface in order to enable testing using a different framework
from the one that we are using now. Now that the methods are not exposed
in the public interface, the lint doesn't apply, so we might as well get
rid of the allows.

Signed-off-by: mulhern <amulhern@redhat.com>