Move idempotent return type to pool level
Everywhere else in the code, we return idempotent data types at the
layer right below the D-Bus layer (such as Pool, Engine, etc.). This moves
the idempotent data type for renaming actions up one level from the backstore
to make it consistent with all of the other return types.

It does not change the Engine interface and is merely an internal change
to make it clearer where idempotence is handled. It also makes the
backstore API consistent.