Consolidate find_all and get_stratis_block_devices
The advantages of this are:
* Devices are only iterated through once, rather than being collected and
re-iterated through. Note that this slightly changes semantics. If there were
some block devices identified as Stratis devices by the udev enumerator,
but these devices did not have stratis metadata or devnodes, the fallback
path will be pursued. Previously, it would have been skipped. The
justification for this change is that that would be a very wierd and
unusual situation, so it doesn't hurt to follow the fallback path in that
case.
* The fallback path is easier to understand.
* If the fallback path is followed, only one attempt is made to read the
Stratis header from each block device found, which is better than two
attempts for some.
* There are now some useful warning messages and these messages are
appropriate for the context in which they are emitted.
* Gets rid of stratis_enumerator method; it made the code a bit busier
without any real advantage
* Logs at the warn level instead of just returning with an error from the
find_all method if there is an error reading the Stratis header on the
standard path.

Signed-off-by: mulhern <amulhern@redhat.com>