Consolidate find_all and get_stratis_block_devices
This allows just iterating over the devices, instead of collecting them
and then reiterating. The other advantage is that the log entries can
be made to match the specific situation better than if the two methods
were separated, as what is worth a warning in one context is not in
another.

Signed-off-by: mulhern <amulhern@redhat.com>