Require at least one blockdev for initialization actions
Currently we do not check that initialization actions like
creating a pool or initializing the cache are provided at least one
blockdev. Given that blockdevs are required for these actions,
this PR adds a check at the top level of the engine to return a
helpful error message if no blockdevs are given.