Do not request non-determinism from the simulator unless explicitly using it
Non-determinism was only implemented in the simulator in a few places.
Given that the simulator is unlikely to become non-deterministic in more
places than it is already, only request non-determinism when associated tests
explicitly use it, by, e.g., setting up a different set of assertions
depending on whether an operation succeeds or fails. Otherwise, assume that
the test is intended to be deterministic.

Remove a setUp method made redundant by removal of call to
ConfigureSimulator.

Signed-off-by: mulhern <amulhern@redhat.com>