Require _DEVICE_STRATEGY to always create at least 1 device
Otherwise, testCreateDifferentBlockdevs may randomly fail if setUp
created a pool w/ 0 blockdevs and then the second CreatePool command,
which should cause a conflict, also randomly uses 0 blockdevs to create the
pool. In that case, even though the two lists of devices are random,
since they are both empty, the devices do not conflict and the second
CreatePool call succeeds, causing the test to fail.

Signed-off-by: mulhern <amulhern@redhat.com>