Make separate local functions for non-udev device identification
Document their rationale as clearly as possible. Note that this is a
behavior change, since the method no longer exits immediately on failure
to open a file or failure to obtain a device number.

This means that the method can only failure on an initual udev error, so
document this in every way possible.

Signed-off-by: mulhern <amulhern@redhat.com>