Remove pool_uuid from the parameters of get_metadata
Strengthen the precondition. In the only context in which get_metadata is
called, all the devices have had their static headers read, and been
identified as belonging to the same pool. Return an error if there is no BDA.

Use collect() to get the bdas rather than a loop; it is clearer.

Signed-off-by: mulhern <amulhern@redhat.com>