Introduce a new method, identify_block_device, in the identify module
Use it in block_evaluate and discard any newly introduced dead code.
Have block_evaluate take a udev device as an argument.

This makes the udev lookup constant time instead of linear in the number
of block devices. Previously, the device node was discovered from the
udev database entry, and then the entries for all block devices in the udev
database were searched in order to locate a block device that had the
specified device node in order to locate the relevant udev database entry.

Simplify block_evaluate's return type. It never returns an error, logging any
internal errors instead.

Signed-off-by: mulhern <amulhern@redhat.com>