Adjust dbus-api names and permissions
Make methods and props modules private in their respective parent modules.

Instead of using version-specific method names, distinguish api methods by
the parent of their api module.  This eliminates a kind of redundancy,
where the method name and the interface name both include the version
number, and that number is always the same.

By specifying an api method by its parent module name, the
definition of the objects and the interfaces they support, always in
mod.rs, become more uniform, and a bit clearer. So, when a second version
of the same interface is added, it is easier to pick out precisely the
origin of each api method, i.e., whether it is an override of a method
from a previous interface or not.

The methods in methods.rs and props.rs are used only to support the
methods in api.rs. So, if these modules are made private, then any methods
that are truly shared between interfaces won't find themselves in these
methods, but instead in the appropriate shared.rs.

Signed-off-by: mulhern <amulhern@redhat.com>