Move udev module back down to backstore where it belongs
It had to be moved up a level because the udev functionality was being
used in strat_engine/engine.rs to decide the ownership in the
block_evaluate method. Now that functionality has been unified into
identify_block_device in the identify module in backstore, so the udev
module can be moved down again and be the identify module's sibling.

Signed-off-by: mulhern <amulhern@redhat.com>