Rewrite initialize function
Remove device module entirely. It only contains DevOwnership, which
is now entirely unused.

The contributions of this rewrite are:
* Decompose the problem better. There are now two functions, one which
gets information about each device and another which initializes the devices.
Previously all this was in one big method, initialize. There are two
advantages:
  * The separated functions are a lot clearer.
  * The interface will allow the management of idempotency to be hoisted up
  the call chain to the engine. Currently, the idempotency treatment on
  create_pool or add must work by analyzing what happened, this will allow it
  to be much more direct and stable. In this commit, the check is moved up
  only one level, to blockdevmgr and abstracted into the function
  check_device_ids.
* It removes the DevOwnership enumerated type entirely. This type had been
designed to be used both when deciding identity to intialize devices and
when deciding whether a device should be included in a pool. This dual
purpose made it tough to use in either context, and it has already been
dropped from the identify module in a previous PR. Tests that made use of
DevOwnership were replaced with tests that made use of UdevOwnership.
* Make resolve_devices private, preparatory to removing it in a subsequent
commit.
* Improve error messages to provide context for various failures.
* Be less eager about performing relatively costly operations that get
results that may not be needed. For example, if udev indicates that the
device is owned by another, there is no need to get its size.
* Adds a FIXME marking the spot where we ought to verify that the device
is truly unowned using libblkid.
* Do not multiply file descriptors gratuitously. Previously, while one file
descriptor was being held on a device, another would be made in order to
read Stratis metadata. Now only one file descriptor is open at a time on
any given device.

Signed-off-by: mulhern <amulhern@redhat.com>