Make a new FetchProperties interface
Make it for all, even though only the pool is getting a new property in
this PR. We must use one global version for the FetchProperties interface,
since it is the same interface for pools, filesystems, and blockdevs.

The number indicates that a different set of properites is supported, not
that the function interface has changed.

Signed-off-by: mulhern <amulhern@redhat.com>