No longer implement Error::description in StratisError
It is substantially the same as the the Display implementation and it is
now hard-deprecated.

It is used by the D-Bus translation function, change that use to
to_string(), i.e., use the Display implementation instead.

We do not support the Rust API of libstratis, so the function can be
removed, rather than just deprecated.

Signed-off-by: mulhern <amulhern@redhat.com>