Add additional comments to Pool::add_blockdevs
It gets to change its semantics with this PR.

Signed-off-by: mulhern <amulhern@redhat.com>

Tweak the code a bit

I like to make the if-expression be constrained to be precisely about the
thing that is dependent on the if-condition.

Signed-off-by: mulhern <amulhern@redhat.com>

Remove encryption work in setup methods

This is a procedural change. I'ld rather handle setup of encrypted device
in an entirely separate PR.

Signed-off-by: mulhern <amulhern@redhat.com>

Add some clarifying comments to various dbus_api methods

Due to the necessity to support multiple versions, and efforts to make
the specification of the Pool trait belonging to the engine more
straightforward, some of these methods need to dispatch in complicated
ways.

Signed-off-by: mulhern <amulhern@redhat.com>

Expand an error message

It's a bit clearer.

Signed-off-by: mulhern <amulhern@redhat.com>

Add a clarifying and justifying comment

Signed-off-by: mulhern <amulhern@redhat.com>

Prohibit cache initialization on an encrypted pool

It's our original design and it keeps things simpler.

Signed-off-by: mulhern <amulhern@redhat.com>

Add encryption and cache mutually exclusive to pool invariant

Signed-off-by: mulhern <amulhern@redhat.com>