Verify that identifying metadata has been written to devices in a test
This test has intermittent failures; it fails to find the devices it should
using find_all. find_all() works, but it uses udev, and the udev database
does not seem to be properly updated, although it should be since we
use udev settle. Verify that the proper identifiers are on the devices
immediately after they are added to the backstore.

Signed-off-by: mulhern <amulhern@redhat.com>