Require libcryptsetup bindings
We will start using them in distinct PRs, we might as well require them
immediately in this PR. That way the associated Cargo.lock changes, which
are somewhat inconvenient to reconcile, can go in immediately.

Also, update Travis CI to support compilation of the libcryptsetup
bindings. These require cryptsetup 2.2.0, so boost distribution to the most
recent available, which is bionic, and install cryptsetup from ancillary
repository specified using before_install.

Signed-off-by: mulhern <amulhern@redhat.com>