Update to later versions of packages
When libcryptsetup-rs began using a later version of bindgen, this
made it so our version of regex in Cargo.lock will no longer be
compatible with the version of bindgen that we need. This PR
provides an updated version of Cargo.lock that will be compatible with
the next version of libcryptsetup-rs released.