Add crypt module and use for device initialization
This commit adds the lower level implementation of encryption on
device implementation. It adds a crypt module which uses
libcryptsetup-rs for the implementation of the encryption operations and
exposes an encryption API to the rest of stratisd. It also provides an
initial implementation of encryption capabilities for stratisd using the
crypt module that is currently disabled at the D-Bus layer. The ability
to invoke the encryption implementation from D-Bus will be added in a
later commit. This commit also provides encryption testing to verify
that the provided devices are actually being encrypted and cleaned up
properly using the crypt module.