Have _expected_stratis_devices take only one parameter
The second was redundant, and really only used to check that there was
no bug in the test code and that Python arithmetic works.

Retain the devices argument rather than the number argument, because there
is sometimes a need to compute the devices for other purposes and because
at sometime it might make sense to make the method more precise, and have
it check the actual devices, rather than just the number.

Signed-off-by: mulhern <amulhern@redhat.com>